rename file to private
