+++ src/miniFasta/__init__.py
@@ -1,4 +1,4 @@
-from .miniFasta import fasta_object, read, write, print_fasta, translate_seq, reverse_comp
+from ._miniFasta import fasta_object, read, write, print_fasta, translate_seq, reverse_comp
 
 __all__ = [
     "fasta_object",
