+++ src/miniFasta/_miniFasta.py
